Fix the desktop-mode problem a better way
I used to have problems with desktop-mode as well, but finally tracked it down (I ran across this page while searching the internet).  I've been using a newer version in which the bug is fixed, but then ran into it again.  The problem is that `desktop-file-version` is an integer 206 which gets written to the file as a character.   In new versions, they changed it to be a string so that it gets written correctly.  Anyway, hope that helps.